Factor out common project-logic between build and check.
