+++ crates/cli/src/cmd/tx/simulate.rs
@@ -121,5 +121,5 @@ pub fn exec(
         }
     }
 
-    return Ok(());
+    Ok(())
 }
