+++ crates/cli/src/cmd/uplc/flat.rs
@@ -79,5 +79,5 @@ pub fn exec(
         }
     }
 
-    return Ok(());
+    Ok(())
 }
